% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beginr-package.R
\docType{package}
\name{beginr-package}
\alias{beginr}
\alias{beginr-package}
\title{Make life easier for beginnErS}
\description{
This is a package that Matt wrote to provide data and functions for R for beginners
}
\details{
ldt_data is data in wide form, see ?ldt_data

dat_long is data in long form, see ?dat_long
}
\keyword{internal}
