% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Butterfly.R
\docType{data}
\name{Butterfly}
\alias{Butterfly}
\alias{Simulation}
\alias{Population}
\alias{World}
\alias{Butterfly}
\alias{Population}
\alias{World}
\title{Example of individual based model using R6: the Butterfly example}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Simulation

Butterfly

Population

World
}
\description{
This example is a simple population genetic model programmed using individual based simulation.
The inspiration is the example of the Peppered moth textbook story (\url{https://en.wikipedia.org/wiki/Peppered_moth_evolution}).
Check examples for usage.
}
\examples{
simu <- Simulation$new(years = 200, N = 500, mutation_rate = 1e-2, period = 50)  ## intialisation
simu$output
simu$run() ## run simulations
simu$plot() ## plot simulations
\dontrun{
simu$profile() ## profile the code
}
}
\keyword{datasets}
