% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_long.R
\docType{data}
\name{dat_long}
\alias{dat_long}
\title{Lexical Decision Task Example Data in long form}
\format{
A data frame with 200 rows and 6 variables:
\describe{
   \item{id}{Participant ID}
   \item{age}{Age of participant in years}
   \item{language}{Languages (1 = monolingual, 2 = bilingual)}
   \item{condition}{Word or Non-Word, a factor}
   \item{rt}{Mean reaction time (ms)}
   \item{acc}{Mean accuracy}
}
}
\source{
<https://github.com/PsyTeachR/introdataviz/tree/master/data-raw>
}
\usage{
dat_long
}
\description{
Simulated Lexical Decision Task data for monolingual and bilingual participants. DVs are mean reaction time (rt) and accuracy (acc) for words and non-words.
}
\keyword{datasets}
