% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deaths_UK}
\alias{deaths_UK}
\title{The deaths census for England and Wales (1901-2016)}
\format{A \var{dataframe} 1445659 obs. of 8 variables.}
\source{
These data were derived files freely available from the Office for National Statistics (\url{http://webarchive.nationalarchives.gov.uk}).
}
\usage{
deaths_UK
}
\description{
This dataset contains numbers and causes of death for England and Wales over the period 1901-2016.
The causes of deaths have been classified using the International Classification of Diseases (ICD).
}
\note{
These data have been prepared for use in the paper "The transition to modernity and chronic disease: mismatch and natural selection" by Stephen Corbett, Alexandre Courtiol, Virpi Lummaa, Jacob Moorad & Stephen Stearns published in Nature Reviews Genetics (2018): \url{https://www.nature.com/articles/s41576-018-0012-3}.
}
\examples{
head(deaths_UK)
}
\keyword{datasets}
