% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo.R
\name{dfplot2}
\alias{dfplot2}
\title{Plot a dataframe, one y against multiple xs}
\usage{
dfplot2(x, y, xlab = "x", ylab = "y", xlim = NULL, ylim = NULL, mycol = NULL, 
    mylty = NULL, xerror = NULL, yerror = NULL, mycolerrorbar = NULL, mylegend = NULL)
}
\arguments{
\item{x}{a vector  or a dataframe with the same length as x}

\item{y}{a vector}

\item{xlab}{character}

\item{ylab}{character}

\item{mycol}{colours}

\item{xerror}{errorbar, same dimension of x}

\item{yerror}{same dimension of y}

\item{mycolerrorbar}{error bar colours}
}
\value{
a figure
}
\description{
Plot a dataframe, one y against multiple xs
}
\examples{

}
