% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beginr.R
\name{errorbar}
\alias{errorbar}
\title{Display errorbars}
\usage{
errorbar(x, y, xupper = NULL, xlower = NULL, yupper = NULL,
  ylower = NULL, col = "black", lty = 1)
}
\arguments{
\item{x, y}{numeric vectors, the coordinates of points at which the error bar
are to be plotted.}

\item{xupper, xlower}{numeric vectors, error value of X axis, the X axis of
error bars are from \eqn{x - xlower} to \eqn{x + xupper}.}

\item{yupper, ylower}{numeric vectors, error value of Y axis, the Y axis of
error bars are from \eqn{y - ylower} to \eqn{y + yupper}.}

\item{col, lty}{graphical parameters, color and line type for the lines, more
details see \code{\link[graphics]{par}}.}
}
\description{
Create a errorbar, or add errorbars to an existing plot.
}
\examples{
x <- seq(0, 2 * pi, length.out = 100)
y <- sin(x)
plot(x, y, type = 'l')
errorbar(x, y, yupper = 0.1, ylower = 0.1)

}
