% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{get_vignettes}
\alias{get_vignettes}
\title{Access the vignettes}
\usage{
get_vignettes()
}
\description{
This function allows to access to the vignettes directly in their folder.
}
\examples{
get_vignettes()

}
