% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iris_long}
\alias{iris_long}
\alias{iris_wide}
\title{Toy datasets based on the iris dataset}
\format{Objects of class \var{dataframe}.}
\source{
These data were derived from \code{iris}
}
\description{
These datasets are subset of the \code{\link[datasets]{iris}} dataset that have been designed to illustrate the use of the functions
\code{\link[tidyr]{spread}} and \code{\link[tidyr]{gather}} from \pkg{\link[tidyr]{tidyr}}.
}
\examples{
iris_long
iris_wide
}
\keyword{datasets}
