% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beginr.R
\name{lmdf}
\alias{lmdf}
\title{calculate linear regression between every two columns in a data frame. in: a dataframes. out: a dataframe showing the linear regression.}
\usage{
lmdf(data, simply = FALSE, intercept = TRUE)
}
\arguments{
\item{data}{a dataframe}

\item{simply}{logical}

\item{intercept}{logical}
}
\value{
another dataframe
}
\description{
calculate linear regression between every two columns in a data frame. in: a dataframes. out: a dataframe showing the linear regression.
}
\examples{
df <- data.frame(a = 1:10, b = 1:10 + rnorm(10), c = 1:10 + rnorm(10))
lmdf(df)

}
