% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Butterfly.R
\name{load_butterfly_example}
\alias{load_butterfly_example}
\title{Example of individual based model using R6: the Butterfly example}
\usage{
load_butterfly_example()
}
\value{
Nothing but directly export created R6 object into the Global Environment
}
\description{
This example is a simple population genetic model programmed using individual based simulation.
The inspiration is the example of the Peppered moth textbook story (\url{https://en.wikipedia.org/wiki/Peppered_moth_evolution}).
}
\examples{
load_butterfly_example()  ## load all the objects
simu <- Simulation$new(years = 200, N = 500, mutation_rate = 1e-2, period = 50)  ## intialisation
simu$output
simu$run() ## run simulations
simu$plot() ## plot simulations
\dontrun{
simu$profile() ## profile the code
}
}
