% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manufacture_ldt_long_data.R
\name{manufacture_ldt_long_data}
\alias{manufacture_ldt_long_data}
\title{Generate long form lexical decision data}
\usage{
manufacture_ldt_long_data()
}
\value{
A rectangular dataset with class tibble
}
\description{
Generate long form lexical decision data
}
\examples{
dat_long = manufacture_ldt_long_data()
str(dat_long)
summary(dat_long)
require("ggplot2")
ggplot(dat_long, aes(x = rt, fill = condition)) +
  geom_density(alpha = 0.75)+
  scale_x_continuous(name = "Reaction time (ms)")+
  scale_fill_discrete(name = "Condition")
}
