% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo.R
\name{mf_bib}
\alias{mf_bib}
\title{Create a bib file for R packages, including the citations of user-defined packages.}
\usage{
mf_bib(pkg = c("base"), bibfile = "packages.bib")
}
\arguments{
\item{pkg}{packages}

\item{bibfile}{file path and name to save the bib entries.}
}
\value{
a bib file
}
\description{
Create a bib file for R packages, including the citations of user-defined packages.
}
\examples{
mf_bib()
mf_bib(pkg = c("base", "knitr"))
}
