% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo.R
\name{mf_dfplot}
\alias{mf_dfplot}
\title{Plot a dataframe, multiple ys against one x}
\usage{
mf_dfplot(x, y, add = FALSE, xlab = "", ylab = "", myaxes = FALSE, xlim = NULL, 
    ylim = NULL, mycol = NULL, mytype = "l", mypch = 20, mycex = 1, mylty = NULL, 
    lwd = 1, xerror = NULL, yerror = NULL, mycolerrorbar = NULL, mylegend = NULL, 
    mylegendcol = mycol, mylegendcex = 1, legendpos = "top")
}
\arguments{
\item{x}{a vector}

\item{y}{a vector or a dataframe with the same length as x}

\item{add}{logical, whether to add this plot to the previous one}

\item{xlab}{character}

\item{ylab}{character}

\item{myaxes}{logical, whether to display axes automatically}

\item{mycol}{colours}

\item{xerror}{errorbar, same dimension of x}

\item{yerror}{same dimension of y}

\item{mycolerrorbar}{error bar colours}

\item{legendpos}{character}

\item{myledendcex}{numeric}
}
\value{
a figure
}
\description{
Plot a dataframe, multiple ys against one x
}
\examples{

}
