% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo.R
\name{mf_errorbar}
\alias{mf_errorbar}
\title{add error bars to a scatterplot.}
\usage{
mf_errorbar(x, y, xupper = NULL, xlower = NULL, yupper = NULL, ylower = NULL, 
    col = "black", lty = 1)
}
\value{
errorbars in a plot
}
\description{
add error bars to a scatterplot.
}
\examples{

}
