% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo.R
\name{mf_hist}
\alias{mf_hist}
\title{Plot a user-customized hist}
\usage{
mf_hist(data, mybreaks = "Sturges", myxlim = NULL, myylim = NULL, eightlines = TRUE, 
    eightdigit = 0, eightcex = 0.8, eightcolors = c("red", "darkgreen", "blue", "black", 
        "purple", "gold")[c(1, 2, 3, 2, 1, 6, 6, 5, 4, 5)], mylegend = "", myxlab = "", 
    show_n = TRUE, show_skewness = TRUE, show_density = FALSE)
}
\value{
a hist plot
}
\description{
Plot a user-customized hist
}
\examples{

}
