% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo.R
\name{mf_list2ascii}
\alias{mf_list2ascii}
\title{Save a list into an ASCII file. in: a list. out: a file.}
\usage{
mf_list2ascii(x, file = paste(deparse(substitute(x)), ".txt", sep = ""))
}
\arguments{
\item{x}{a list}
}
\value{
a file
}
\description{
Save a list into an ASCII file. in: a list. out: a file.
}
\examples{

}
