% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo.R
\name{mf_lmdf}
\alias{mf_lmdf}
\title{calculate linear regression between every two columns in a data frame. in: a dataframes. out: a dataframe showing the linear regression.}
\usage{
mf_lmdf(data, simply = FALSE, intercept = TRUE)
}
\arguments{
\item{data}{a datafrram}
}
\value{
another dataframe
}
\description{
calculate linear regression between every two columns in a data frame. in: a dataframes. out: a dataframe showing the linear regression.
}
\examples{

}
