% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo.R
\name{mf_readdir}
\alias{mf_readdir}
\title{batch read several files. in: a path. out: a list containing all data frames}
\usage{
mf_readdir(wd = ".", sep = c(","))
}
\arguments{
\item{wd}{working dir}

\item{sep}{sep of each file}
}
\value{
a list
}
\description{
batch read several files. in: a path. out: a list containing all data frames
}
\examples{

}
