% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo.R
\name{mf_skewness}
\alias{mf_skewness}
\title{Calculate the skewness of a distribution}
\usage{
mf_skewness(x)
}
\arguments{
\item{x}{the data to check}
}
\value{
the skewness of the distribution of x
}
\description{
Calculate the skewness of a distribution
}
\examples{
mf_skewness(rnorm(100))
}
