% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo.R
\name{mf_tapply}
\alias{mf_tapply}
\title{a friendly version of tapply for dataframes. in and out: same as tapply(). the built-in function tapply returns a matrix with unfriendly row name and colname. mf-tapply returns a friendly dataframe}
\usage{
mf_tapply(data, select = names(data), myfactor, ..., na.rm = c(TRUE, FALSE, NULL)[1])
}
\arguments{
\item{data}{dataframe}

\item{select}{character, column names to calc}

\item{myfactor}{a colname as factor}
}
\value{
a dataframe
}
\description{
a friendly version of tapply for dataframes. in and out: same as tapply(). the built-in function tapply returns a matrix with unfriendly row name and colname. mf-tapply returns a friendly dataframe
}
\examples{

}
