% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo.R
\name{mf_write}
\alias{mf_write}
\title{save csv file with asking if the file already exists.}
\usage{
mf_write(data, writefile, row.names = FALSE)
}
\arguments{
\item{writefile}{destination file}

\item{row.names}{logical}
}
\value{
write a file
}
\description{
save csv file with asking if the file already exists.
}
\examples{

}
