% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo.R
\name{name}
\alias{name}
\title{Enhancement of names()}
\usage{
name(data)
}
\arguments{
\item{data}{a dataframe}
}
\value{
a list
}
\description{
Enhancement of names()
}
\examples{
df <- data.frame(a = NA, b = NA, c = NA)
name(df)

}
