% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packr.R
\name{packr}
\alias{packr}
\title{Create a package}
\usage{
packr(pkg_name, packages, author = NULL, email = NULL, auto = FALSE, overwrite = FALSE)
}
\arguments{
\item{pkg_name}{the name of the package which is to be created}

\item{packages}{packages wrapped in this group}

\item{author}{author of the new package}

\item{email}{email of the author}

\item{auto}{logical. whether to build and install the new package automatically}

\item{overwrite}{logical. whether to overwrite the package with the same name if it already installed}
}
\value{
a folder with a package skeleton
}
\description{
Create a package
}
\examples{
\dontrun{
packr("zhaor", c("mindr", "pinyin", "beginr", "bookdownplus", "steemr", "rmd"), 
    "Your Name")
}
}
