% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beginr.R
\name{plotcolorbar}
\alias{plotcolorbar}
\title{A reminder for color bars. More palettes can be found in 'colormap', 'RColorBrewer', and 'dichromat' packages.}
\usage{
plotcolorbar()
}
\value{
a figure
}
\description{
A reminder for color bars. More palettes can be found in 'colormap', 'RColorBrewer', and 'dichromat' packages.
}
\examples{
plotcolorbar()
}
