% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beginr.R
\name{plotlm}
\alias{plotlm}
\title{plot a linear regression figure and return a list of parameters.}
\usage{
plotlm(x, y, xlim = range(as.numeric(x), na.rm = TRUE),
  ylim = range(as.numeric(y), na.rm = TRUE),
  plot.title = "linear regression", xlab = "x", ylab = "y",
  refline = FALSE, slope = 1, intercept = 0, showr2 = TRUE,
  showleg = TRUE)
}
\arguments{
\item{x}{numeric}

\item{y}{numeric}

\item{xlim}{numeric}

\item{ylim}{numeric}

\item{plot.title}{character}

\item{xlab}{character}

\item{ylab}{character}

\item{refline}{logical. if a reference line is plotted}

\item{slope}{slope of refline}

\item{intercept}{intercept of refline}

\item{showr2}{logical}

\item{showleg}{logical}
}
\value{
a figure
}
\description{
plot a linear regression figure and return a list of parameters.
}
\examples{
plotlm(1:10, 1:10 + rnorm(10))

}
