% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beginr.R
\name{plotlty}
\alias{plotlty}
\title{A reminder for lty}
\usage{
plotlty(mylwd = 1)
}
\arguments{
\item{mylwd}{numeric. line width}
}
\value{
a figure reminding you lty
}
\description{
A reminder for lty
}
\examples{
plotlty()

}
