% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beginr.R
\name{plotpkg}
\alias{plotpkg}
\title{plot daily download counts of packages}
\usage{
plotpkg(mypkg = "bookdownplus", from = Sys.Date() - 30, to = Sys.Date(), type = "o", 
    pch = 19, col = "blue", cex = 1, textcex = 5)
}
\arguments{
\item{mypkg}{character vector of package names.}

\item{from}{character in 'Y-m-d'}

\item{to}{character in 'Y-m-d'}

\item{type}{the same as that in 'plot()'}

\item{pch}{the same as that in 'plot()'}

\item{col}{the same as that in 'plot()'}

\item{cex}{the same as that in 'plot()'}

\item{textcex}{cex of the package name}
}
\value{
a figure
}
\description{
plot daily download counts of packages
}
\examples{
plotpkg(mypkg = "rmarkdown")
}
