% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population_UK}
\alias{population_UK}
\title{The population census for England and Wales (1901-2016)}
\format{A \var{dataframe} with 1392 obs. of 4 variables.}
\source{
These data were derived files freely available from the Office for National Statistics (\url{http://webarchive.nationalarchives.gov.uk}).
}
\description{
This dataset contains the total population count (per age classes) for England and Wales over the period 1901-2016.
}
\note{
These data have been prepared for use in the paper "The transition to modernity and chronic disease: mismatch and natural selection" by Stephen Corbett, Alexandre Courtiol, Virpi Lummaa, Jacob Moorad & Stephen Stearns published in Nature Reviews Genetics (2018): \url{https://www.nature.com/articles/s41576-018-0012-3}.
}
\examples{
head(population_UK)
}
\keyword{datasets}
