% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo.R
\name{readdir}
\alias{readdir}
\title{Read multiple tables into a list.}
\usage{
readdir(mydir = getwd(), sep = c(","))
}
\arguments{
\item{mydir}{the folder path}

\item{sep}{sep of each file}
}
\value{
a list
}
\description{
Read multiple tables into a list.
}
