% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packr.R
\name{rplc}
\alias{rplc}
\title{Replace strings in a file}
\usage{
rplc(oldchar, newchar, filename)
}
\arguments{
\item{oldchar}{old string}

\item{newchar}{new string}

\item{filename}{file name}
}
\value{
modified files
}
\description{
Replace strings in a file
}
