% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beginr.R
\name{tapply2}
\alias{tapply2}
\title{a friendly version of tapply for a column in a dataframe}
\usage{
tapply2(data, select = names(data)[1], myfactor, ..., na.rm = c(TRUE, FALSE, NULL)[1])
}
\arguments{
\item{data}{dataframe}

\item{select}{character, column names to calc}

\item{myfactor}{a colname as factor}

\item{...}{function to apply to data}

\item{na.rm}{logical}
}
\value{
a dataframe
}
\description{
a friendly version of tapply for a column in a dataframe
}
