% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beginr.R
\name{tapplydf}
\alias{tapplydf}
\title{a friendly version of tapply for dataframes}
\usage{
tapplydf(data, select = names(data), myfactor, ..., na.rm = c(TRUE,
  FALSE, NULL)[1])
}
\arguments{
\item{data}{dataframe}

\item{select}{character, column names to calc}

\item{myfactor}{a colname as factor}

\item{...}{function to apply to data}

\item{na.rm}{logical}
}
\value{
a dataframe
}
\description{
a friendly version of tapply for dataframes
}
