% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beginr.R
\name{tapplydfv}
\alias{tapplydfv}
\title{a friendly version of tapply}
\usage{
tapplydfv(colname = "tapply", x, factor, ...)
}
\arguments{
\item{colname}{character}

\item{x}{a datafrom}

\item{factor}{factor for tapply}

\item{...}{the function to apply to data}
}
\value{
a dataframe
}
\description{
a friendly version of tapply
}
