% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beginr.R
\name{writefile}
\alias{writefile}
\title{save csv file with asking if the file already exists.}
\usage{
writefile(data, writefile, row.names = FALSE)
}
\arguments{
\item{data}{a data frame}

\item{writefile}{destination file}

\item{row.names}{logical}
}
\value{
write a file
}
\description{
save csv file with asking if the file already exists.
}
