% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beginr-package.R
\docType{package}
\name{beginr-package}
\alias{beginr}
\alias{beginr-package}
\title{Make life easier for beginnErS}
\description{
`manufacture_ldt_long_data` creates the example data that we use in the _R for beginners_ course
}
\details{
This is a package that Matt wrote to provide data and functions

See ?manufacture_ldt_long_data

More soon
}
\keyword{internal}
