% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beginr.R
\name{bib}
\alias{bib}
\title{Create the bibtex entry}
\usage{
bib(pkgs = "base", bibfile = "")
}
\arguments{
\item{pkgs}{character vector of package names to cite}

\item{bibfile}{character string, file path to save the bib entries.  If ""
(the default), it prints to the standard output connection, or save the bibtex
entry to \code{bibfile}.}
}
\value{
None (invisible NULL)
}
\description{
Output the bibtex entry of installed packages. If \code{bibfile} is not "",
bibentry will be saved to bibfile.
}
\examples{
bib()

\dontrun{
bib(pkg = c('mindr', 'bookdownplus', 'pinyin'))
}
}
