% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beginr.R
\name{bib}
\alias{bib}
\title{Create a bib file for R packages, including the citations of user-defined packages.}
\usage{
bib(pkg = c("base"), bibfile = "")
}
\arguments{
\item{pkg}{character. Packages}

\item{bibfile}{character. File path and name to save the bib entries.  If "" (the default), it prints to the standard output connection, the console unless redirected by sink.}
}
\value{
bib entries
}
\description{
Create a bib file for R packages, including the citations of user-defined packages.
}
\examples{
bib()
bib(pkg = c("mindr", "bookdownplus", "pinyin"))
}
