% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo.R
\name{errorbar}
\alias{errorbar}
\title{add error bars to a scatterplot.}
\usage{
errorbar(x, y, xupper = NULL, xlower = NULL, yupper = NULL, ylower = NULL, 
    col = "black", lty = 1)
}
\value{
errorbars in a plot
}
\description{
add error bars to a scatterplot.
}
\examples{

}
