% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beginr.R
\name{errorbar}
\alias{errorbar}
\title{add error bars to a scatterplot.}
\usage{
errorbar(x, y, xupper = NULL, xlower = NULL, yupper = NULL, ylower = NULL, 
    col = "black", lty = 1)
}
\arguments{
\item{x}{numeric}

\item{y}{numeric}

\item{xupper}{numeric}

\item{xlower}{numeric}

\item{yupper}{numeric}

\item{ylower}{numeric}

\item{col}{colors}

\item{lty}{numeric}
}
\value{
errorbars in a plot
}
\description{
add error bars to a scatterplot.
}
\examples{
x <- seq(0, 2 * pi, length.out = 100)
y <- sin(x)
plot(x, y, type = "l")
errorbar(x, y, yupper = 0.1, ylower = 0.1)

}
