% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beginr.R
\name{list2ascii}
\alias{list2ascii}
\title{Save a list into an ASCII file. in: a list. out: a file.}
\usage{
list2ascii(x, file = paste(deparse(substitute(x)), ".txt", sep = ""))
}
\arguments{
\item{x}{a list}

\item{file}{character. file name}
}
\value{
a file
}
\description{
Save a list into an ASCII file. in: a list. out: a file.
}
\examples{
alist <- list(a = 1:10, b = letters)
list2ascii(alist)

}
