% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo.R
\name{plotlm}
\alias{plotlm}
\title{plot a linear regression figure and return a list of parameters. in: two vectors. out: a figure and a list.}
\usage{
plotlm(x, y, xlim = range(as.numeric(x), na.rm = TRUE), ylim = range(as.numeric(y), 
    na.rm = TRUE), plot.title = "linear regression", xlab = "x", ylab = "y", 
    refline = FALSE, slope = 1, intercept = 0, showr2 = TRUE, showleg = TRUE)
}
\arguments{
\item{refline}{logical, reference line}

\item{slope}{slope of refline}

\item{intercept}{of refline}

\item{showr2}{logical}

\item{showleg}{logical}
}
\value{
a figure
}
\description{
plot a linear regression figure and return a list of parameters. in: two vectors. out: a figure and a list.
}
\examples{

}
