% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beginr.R
\name{plotpairs2}
\alias{plotpairs2}
\title{plot pair-wise correlations  with p value. in: a dataframe. out: a figure.}
\usage{
plotpairs2(data, lower.panel = panel.smooth, upper.panel = panel.cor, 
    diag.panel = panel.diag, lwd = 2, col = "grey", labels = "", cex.labels = 4)
}
\arguments{
\item{data}{a dataframe}

\item{lower.panel}{can be panel.lm or panel.smooth}

\item{upper.panel}{panel.cor}

\item{diag.panel}{panel.diag}

\item{lwd}{numeric}

\item{col}{colors}

\item{labels}{character}

\item{cex.labels}{character}
}
\value{
a pair plot
}
\description{
plot pair-wise correlations  with p value. in: a dataframe. out: a figure.
}
\examples{
df <- data.frame(a = 1:10, b = 1:10 + rnorm(10), c = 1:10 + rnorm(10))
plotpairs2(df)
}
