% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo.R
\name{plotpkg}
\alias{plotpkg}
\title{plot daily download counts of packages}
\usage{
plotpkg(mypkg = c("bookdownplus", "mindr", "pinyin", "beginr")[1], 
    from = c("2017-06-21", "2017-06-19", "2017-06-19", "2017-06-23")[1], 
    to = Sys.Date(), type = "o", pch = 19, col = "blue", cex = 1, textcex = 5)
}
\arguments{
\item{mypkg}{character vector of package names.}

\item{from}{character in 'Y-m-d'}

\item{to}{character in 'Y-m-d'}

\item{type}{the same as that in 'plot()'}

\item{pch}{the same as that in 'plot()'}

\item{col}{the same as that in 'plot()'}

\item{cex}{the same as that in 'plot()'}

\item{textcex}{cex of the package name}
}
\value{
a figure
}
\description{
plot daily download counts of packages
}
\examples{
plotpkg(mypkg = "rmarkdown")
}
