% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo.R
\name{readdir}
\alias{readdir}
\title{Read multiple tables into a list.}
\usage{
readdir(mydir = getwd(), sep = c(","), output = c("list", "data.frame"), header = TRUE, 
    skip = 0)
}
\arguments{
\item{mydir}{the folder path}

\item{sep}{the field separator character.}

\item{output}{the type of the output. 'list' or 'data.frame'.}

\item{header}{logical. Indicating whether the file contains the names of the variables as its first line.}

\item{skip}{the number of lines of the data file to skip before beginning to read data.}
}
\value{
a list or a data frame
}
\description{
Read multiple tables into a list.
}
